# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Listings Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/listings/{id}/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Получение отзывов на объявлении",
                "operationId": "35af10c7e2b7c9119470a7b8bb96daf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Отзывы создан"
                    },
                    "404": {
                        "description": "Объявление не найдено"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Создание отзыва",
                "operationId": "2558421061d384fc4d264bcf784f2cb3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id объявления",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "rate",
                        "in": "query",
                        "description": "Рейтинг от 1 до 5",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "text",
                        "in": "query",
                        "description": "Текст отзыва",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Отзыв создан"
                    },
                    "404": {
                        "description": "Объявление не найдено"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}